requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==4.2
django-nested-admin==4.0.2
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
pillow==10.3.0
promise==2.3
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
